% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fabp_lin_reg.R
\name{fabp_lin_reg}
\alias{fabp_lin_reg}
\title{Compute FAB p-values under linear regression linking model}
\usage{
fabp_lin_reg(
  Y,
  S,
  R,
  U,
  V,
  pool_sampling_var = T,
  Y1 = NULL,
  S1 = NULL,
  R1 = NULL
)
}
\arguments{
\item{Y}{matrix of averages of experimental readout values over R replicates}

\item{S}{matrix of standard errors of experimental readout values over R replicates}

\item{R}{number of replicates per readout value}

\item{U}{the row features}

\item{V}{the column features}

\item{pool_sampling_var}{logical; indicates whether sampling variance should be assumed the same across hypothesis tests}

\item{Y1}{matrix of averages of experimental readout values over R1 replicates (used for contrast scores for two-sample t-tests; Y is taken to be the other sample)}

\item{S1}{matrix of standard errors of experimental readout values in Y1 over R1 replicates}

\item{R1}{number of replicates per readout value in Y1}
}
\value{
A data.frame of FAB p-values and the standard UMP p-values, one for each entry in Y (or each contrast score Y - Y1)
}
\description{
Comutes FAB p-values based on a linear regression linking model for matrix data given row and column covariates.
}
