% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta.R
\name{genDelta}
\alias{genDelta}
\title{generate Matrix D^{k+1}}
\usage{
genDelta(n, k, x)
}
\arguments{
\item{n}{sample size}

\item{k}{order of fit}

\item{x}{vector of inputs on domain}
}
\description{
generate Matrix D^{k+1}
}
\author{
Edward A. Roualdes
}
