% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior.R
\name{getPost}
\alias{getPost}
\title{get posterior draws from btf object}
\usage{
getPost(btf, parameter = c("f", "s2", "lambda", "omega"), burn = 1000)
}
\arguments{
\item{btf}{btf object}

\item{parameter}{name of the parameter of interest}

\item{burn}{number of samples to discard}
}
\description{
Return posterior draws from a btf object into a coda object
}
\seealso{
\code{\link[btf]{getPostEst}}
}
