% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/plot.btf.R
\name{plot.btf}
\alias{plot.btf}
\title{plot btf object}
\usage{
\method{plot}{btf}(x, t = NULL, burn = 1000, ptest = median,
  prob = 0.95, ...)
}
\arguments{
\item{x}{btf object}

\item{t}{domain of function}

\item{burn}{size of burn-in,}

\item{ptest}{function specifying how draws from the posterior are summarized}

\item{prob}{Numeric scalar in (0,1) giving target probability content of the intervals.}

\item{...}{extra arguments}
}
\description{
Plots a btf object with optional credible intervals.
}
\author{
Edward A. Roualdes
}

